+++ docs/conf.py
@@ -36,11 +36,12 @@ import pignacio_scripts
 # -- General configuration ---------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
-#needs_sphinx = '1.0'
+needs_sphinx = '1.3'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.viewcode']
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.viewcode',
+              'sphinx.ext.napoleon']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -55,7 +56,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'My scripts'
+project = u'pignacio\'s scripts'
 copyright = u'2015, Ignacio Rossi'
 
 # The version info for the project you're documenting, acts as replacement
@@ -209,7 +210,7 @@ latex_elements = {
 # [howto/manual]).
 latex_documents = [
     ('index', 'pignacio_scripts.tex',
-     u'My scripts Documentation',
+     u'pignacio\'s scripts Documentation',
      u'Ignacio Rossi', 'manual'),
 ]
 
@@ -240,7 +241,7 @@ latex_documents = [
 # (source start file, name, description, authors, manual section).
 man_pages = [
     ('index', 'pignacio_scripts',
-     u'My scripts Documentation',
+     u'pignacio\'s scripts Documentation',
      [u'Ignacio Rossi'], 1)
 ]
 
@@ -255,7 +256,7 @@ man_pages = [
 #  dir menu entry, description, category)
 texinfo_documents = [
     ('index', 'pignacio_scripts',
-     u'My scripts Documentation',
+     u'pignacio\'s scripts Documentation',
      u'Ignacio Rossi',
      'pignacio_scripts',
      'One line description of project.',