+++ HISTORY.rst
@@ -4,4 +4,12 @@ History
 -------
 
 0.0.1 (in development)
----------------------
+----------------------
+
+* Packaging
+* Namedtuple with defaults
+* Mock namedtuple
+* NagiosLogger
+* Custom TestCase
+* capture_stdout
+* Colors for the terminal